Drop core docs plugin, this is handled by legacy rest api plugin. Add configuration for plugin name/description. Simplify dependencies on legacy rest muck.
