--- nexus/nexus-core-plugins/nexus-core-documentation-plugin/pom.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Sonatype Nexus (TM) Open Source Version
-    Copyright (c) 2007-2012 Sonatype, Inc.
-    All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
-
-    This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
-    which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
-
-    Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
-    of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
-    Eclipse Foundation. All other trademarks are the property of their respective owners.
-
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.sonatype.nexus.plugins</groupId>
-    <artifactId>nexus-core-plugins</artifactId>
-    <version>2.3-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>nexus-core-documentation-plugin</artifactId>
-  <packaging>nexus-plugin</packaging>
-  <name>Nexus : Core Plugins : Core Documentation</name>
-
-  <properties>
-    <nexus.plugin.name>Nexus Core Documentation Plugin</nexus.plugin.name>
-    <nexus.plugin.description>Exposes Nexus core documentation.</nexus.plugin.description>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.sonatype.nexus.plugins.rest</groupId>
-      <artifactId>nexus-legacy-rest-api-plugin</artifactId>
-      <version>${project.version}</version>
-      <scope>provided</scope>
-      <type>${nexus-plugin.type}</type>
-    </dependency>
-
-    <dependency>
-      <groupId>org.sonatype.nexus</groupId>
-      <artifactId>nexus-api</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>org.codehaus.plexus</groupId>
-      <artifactId>plexus-component-annotations</artifactId>
-      <scope>provided</scope>
-    </dependency>
-  </dependencies>
-
-</project>
