+++ nexus/nexus-core-plugins/nexus-legacy-rest-api-parent/nexus-legacy-rest-api-plugin/src/main/java/org/sonatype/nexus/internal/CoreDocumentationResourceBundle.java
@@ -10,44 +10,46 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.plugin.coredocumentation;
+package org.sonatype.nexus.internal;
 
+import org.sonatype.nexus.plugins.rest.AbstractDocumentationNexusResourceBundle;
+import org.sonatype.nexus.rest.NexusApplication;
+
+import javax.inject.Named;
+import javax.inject.Singleton;
 import java.io.IOException;
 import java.util.zip.ZipFile;
 
-import org.codehaus.plexus.component.annotations.Component;
-import org.sonatype.nexus.plugins.rest.AbstractDocumentationNexusResourceBundle;
-import org.sonatype.nexus.plugins.rest.NexusResourceBundle;
-import org.sonatype.security.rest.AbstractSecurityPlexusResource;
-
-@Component( role = NexusResourceBundle.class, hint = "SecurityDocumentationResourceBundle" )
-public class SecurityDocumentationResourceBundle
+/**
+ * @since 2.3
+ */
+@Named
+@Singleton
+public class CoreDocumentationResourceBundle
     extends AbstractDocumentationNexusResourceBundle
 {
 
     @Override
     public String getPluginId()
     {
-        return "nexus-core-documentation-plugin";
-    }
-
-    @Override
-    public String getUrlSnippet()
-    {
-        return "security";
+        return "nexus-legacy-rest-api-plugin";
     }
 
     @Override
     public String getDescription()
     {
-        return "Security API";
+        return "Core (Reslet 1.x) API";
+    }
+
+    @Override
+    public String getPathPrefix() {
+        return "core";
     }
 
     @Override
     protected ZipFile getZipFile()
         throws IOException
     {
-        return getZipFile( AbstractSecurityPlexusResource.class );
+        return getZipFile( NexusApplication.class );
     }
-
 }
