+++ nexus/nexus-core-plugins/nexus-legacy-rest-api-parent/nexus-legacy-rest-api-plugin/src/main/java/org/sonatype/nexus/internal/SecurityDocumentationResourceBundle.java
@@ -10,43 +10,47 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.plugin.coredocumentation;
+package org.sonatype.nexus.internal;
 
+import org.sonatype.nexus.plugins.rest.AbstractDocumentationNexusResourceBundle;
+import org.sonatype.security.rest.AbstractSecurityPlexusResource;
+
+import javax.inject.Named;
+import javax.inject.Singleton;
 import java.io.IOException;
 import java.util.zip.ZipFile;
 
-import org.codehaus.plexus.component.annotations.Component;
-import org.sonatype.nexus.plugins.rest.AbstractDocumentationNexusResourceBundle;
-import org.sonatype.nexus.plugins.rest.NexusResourceBundle;
-import org.sonatype.nexus.rest.NexusApplication;
-
-@Component( role = NexusResourceBundle.class, hint = "CoreDocumentationResourceBundle" )
-public class CoreDocumentationResourceBundle
+/**
+ * @since 2.3
+ */
+@Named
+@Singleton
+public class SecurityDocumentationResourceBundle
     extends AbstractDocumentationNexusResourceBundle
 {
 
     @Override
     public String getPluginId()
     {
-        return "nexus-core-documentation-plugin";
+        return "nexus-legacy-rest-api-plugin";
+    }
+
+    @Override
+    public String getDescription()
+    {
+        return "Security (Restlet 1.x) API";
+    }
+
+    @Override
+    public String getPathPrefix() {
+        return "security";
     }
 
     @Override
     protected ZipFile getZipFile()
         throws IOException
     {
-        return getZipFile( NexusApplication.class );
+        return getZipFile( AbstractSecurityPlexusResource.class );
     }
 
-    @Override
-    public String getDescription()
-    {
-        return "Core API";
-    }
-
-    @Override
-    protected String getUrlSnippet()
-    {
-        return "core";
-    }
 }
