+++ nexus/nexus-core-plugins/nexus-legacy-rest-api-parent/pom.xml
@@ -26,6 +26,18 @@
   <name>Nexus : Core Plugins : Legacy REST API</name>
   <packaging>pom</packaging>
 
+  <dependencyManagement>
+    <dependencies>
+      <!-- INTERNAL -->
+
+      <dependency>
+        <groupId>org.sonatype.nexus.plugins.rest</groupId>
+        <artifactId>nexus-legacy-rest-api-model</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
+
   <modules>
     <module>nexus-legacy-rest-api-model</module>
     <module>nexus-legacy-rest-api-plugin</module>
