+++ nexus/nexus-core-plugins/pom.xml
@@ -45,7 +45,6 @@
     <module>nexus-plugin-console-plugin</module>
     <module>nexus-ldap-plugin-parent</module>
     <module>nexus-rrb-plugin</module>
-    <module>nexus-core-documentation-plugin</module>
     <module>nexus-archetype-plugin</module>
     <module>nexus-maven-bridge-plugin</module>
   </modules>
